Adiciona ações para manipulação de uma loja + testes de integração
