+++ .gitignore
@@ -21,6 +21,5 @@
 
 .idea
 
-docker/.env
-coverage/*
 /.env
+coverage/*
