+++ app/controllers/concerns/error_handler.rb
@@ -1,0 +1,10 @@
+module ErrorHandler
+
+  def build_error(message)
+    {
+      'error': {
+        'message': message
+      }
+    }
+  end
+end
