+++ config/routes.rb
@@ -11,6 +11,7 @@ Rails.application.routes.draw do
     namespace :v1 do
 
       resources :stores
+      
     end
   end
